// src/types/mock-data.type.ts
export type MockData = {
  title: string[];
  description: string[];
  city: string[];
  previewImage: string[];
  images: string[];
  isPremium: boolean[];
  isFavorite: boolean[];
  type: string[];
  goods: string[];
  host: {
    name: string[];
    email: string;
    avatarUrl: string;
    password: string;
    isPro: boolean[];
  }
}